More intelligible display of new jobs.
Lists by date and category.  Right now, the display includes each
category, including the zip codes -- which isn't obviously desirable,
but *does* help illustrate a massive data cleaning problem:  every
single new job for the past week is coded with zip code 02108,
which someone seems to be using as a generic "Boston" zip code.
In fact, places with that address are all within a stone's throw
of the state house, and most of these jobs are pretty far-flung...